chore: ai split
