+++ src/ui/common/components/Multistaking/ChainSelectionModal/shared/index.ts
@@ -1,0 +1,3 @@
+export { ChainButton } from "./ChainButton";
+export { SubSection } from "./SubSection";
+export type { ChainButtonProps, SubSectionProps } from "./types";
