+++ src/ui/common/components/Multistaking/ChainSelectionModal/shared/types.ts
@@ -1,0 +1,19 @@
+import { FinalityProvider } from "@/ui/common/types/finalityProviders";
+
+export interface ChainButtonProps {
+  disabled?: boolean;
+  provider?: FinalityProvider;
+  bsnId?: string;
+  bsnName?: string;
+  logoUrl?: string;
+  title?: string | JSX.Element;
+  onSelectFp?: () => void;
+  onRemove?: (bsnId: string) => void;
+  isExisting?: boolean;
+}
+
+export interface SubSectionProps {
+  children: React.ReactNode;
+  style?: React.CSSProperties;
+  className?: string;
+}
