+++ src/ui/common/components/Multistaking/MultistakingForm/FinalityProvidersSection.tsx
@@ -123,7 +123,6 @@ export function FinalityProvidersSection() {
         onNext={handleNext}
         onClose={handleClose}
         onSelect={handleSelectBsn}
-        onRemove={handleRemove}
       />
 
       <FinalityProviderModal
